refactor: de-interface logic in Cluster Controller
The logic structure remains the same, but the ClusterStatusUpdater and
the ClusterControl interfaces have been removed.
